% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_all.R
\name{comb_all}
\alias{comb_all}
\title{Combinations among all loci.}
\usage{
comb_all(pos)
}
\arguments{
\item{pos}{A numeric vector. Increasingly sorted, contains the positions of loci.}
}
\value{
A numeric matrix.
}
\description{
Combinations among all loci.
}
\examples{
NULL 
}
\author{
dominik
}

