% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_all.R
\name{comb_all_sets}
\alias{comb_all_sets}
\title{Combinations among all loci of two sets.}
\usage{
comb_all_sets(indices_a, indices_b, pos_a, pos_b)
}
\arguments{
\item{indices_a}{A integer vector. Indices of the first set of loci.}

\item{indices_b}{A integer vector. Indices of the second set of loci.}

\item{pos_a}{A numeric vector. Increasingly sorted, contains the positions of the first set of loci.}

\item{pos_b}{A numeric vector. Increasingly sorted, contains the positions of the second set of loci.}
}
\value{
A numeric matrix.
}
\description{
Combinations among all loci of two sets.
}
\examples{
NULL 
}
\author{
dominik
}

