% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_wind.R
\name{comb_wind}
\alias{comb_wind}
\title{Combinations among loci with minimum and maximum distance.}
\usage{
comb_wind(pos, min_dist, max_dist)
}
\arguments{
\item{pos}{A numeric vector. Increasingly sorted, contains the positions of loci.}

\item{min_dist}{A double. Minimum distance.}

\item{max_dist}{A double. Maximum distance.}
}
\value{
A numeric matrix.
}
\description{
Combinations among loci with minimum and maximum distance.
}
\examples{
comb_wind(pos, min_dist = c(0.1, 0.3, 0.8), max_dist = c(0.2, 0.7, 0.9))
}
\author{
Dominik Mueller (\email{dominikmueller64@yahoo.de})
}

