% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_wind.R
\name{comb_wind_sets}
\alias{comb_wind_sets}
\title{Combinations among different sets of loci with minimum and maximum distance.}
\usage{
comb_wind_sets(indices_a, indices_b, pos_a, pos_b, min_dist, max_dist)
}
\arguments{
\item{indices_a}{A integer vector. Indices of the first set of loci.}

\item{indices_b}{A integer vector. Indices of the second set of loci.}

\item{pos_a}{A numeric vector. Increasingly sorted, contains the positions of the first set of loci.}

\item{pos_b}{A numeric vector. Increasingly sorted, contains the positions of the second set of loci.}

\item{min_dist}{A double. Minimum distance.}

\item{max_dist}{A double. Maximum distance.}
}
\value{
A numeric matrix.
}
\description{
Combinations among different sets of loci with minimum and maximum distance.
}
\examples{
NULL
}
\author{
Dominik Mueller (\email{dominikmueller64@yahoo.de})
}

