% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_geno.R
\name{gen_geno}
\alias{gen_geno}
\title{Generate unphased genotypes}
\usage{
gen_geno(nucleotides = c("A", "C", "G", "T"), n, m, prob_na = 0.2,
  shape1 = 4, shape2 = 4, sep = "/")
}
\arguments{
\item{nucleotides}{A character vector. The allowed nucleatides.}

\item{n}{An integer. The number of genotypes.}

\item{m}{An integer. The number of loci.}

\item{prob_na}{A numeric. The probability of a missing value.}

\item{shape1}{A numeric. The first shape parameter of the beta distribution.}

\item{shape2}{A numeric. The second shape parameter of the beta distribution.}

\item{sep}{A character. The seperator used between nucleotides.}
}
\description{
Generate a random sample of unphased genotypes.
}
\details{
This function can be used to simulate genotypes and is mainly for
testing purposes. 
#' Note: If \code{shape1 == shape2}, the beta distribution is symmetric.
}
\examples{
NULL
 
}
\author{
Dominik Mueller (\email{dominikmueller64@yahoo.de})
}

