% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LD}
\alias{LD}
\title{Compute linkage disequilibrium (LD)}
\usage{
LD(x, y, is_phased = TRUE, any_na = TRUE, r_only = FALSE, check = TRUE)
}
\arguments{
\item{x}{A numeric vector. The haplotype/genotype data of the first locus. If they are phased,
there must only contain 0 and 1. If they are unphased, they must only contain
0, 1, or 2, where 1 codes for the heterozygot.}

\item{y}{A numeric vector. The haplotype/genotype data of the second locus.}

\item{is_phased}{A logical. Are the data phased?}

\item{any_na}{A logical. May some genotypes contain missing values? If not,
computations are more efficient for phased genotypes.}

\item{r_only}{Should only the r-statistic be computed?
(Produces minimal output for saving memory and computation time.)}

\item{check}{A logical. Should checks be performed?}
}
\value{
\code{LD} returns a list with the following components:
\itemize{
  \item{D}{ Linkage disequilibrium estimate}
  \item{Dprime}{ Scaled linkage disequilibrium estimate}
  \item{r}{ Correlation coefficient} 
  \item{r2}{ Squared correlation coefficient} 
  \item{n}{ Number of observations}
  \item{chis2}{ Chi-square statistic for linkage
    equilibrium (i.e., D = Dprime = r = r2 = 0)}
  \item{pval}{ Chi-square p-value for marker independence}
}
}
\description{
Compute the linkage disequilibrium (LD) between a pair of loci.
}
\details{
For computing LD among a large number of loci, please use 
\code{\link{LD_mult}}, which is much more efficient for that.

Linkage disequilibrium (LD) is the non-random association of
marker alleles and can arise from marker proximity or from selection
bias.
 
Three estimators of LD are computed:
 \itemize{
   
   \item{D}{ raw difference in frequency between the
      observed number of AB pairs and the expected number:
        
        \deqn{%
          D = p_{AB} - p_A p_B %
        }{%
          D = p(AB) - p(A)*p(B) %
        }
      
    }
    \item{D'}{ scaled D spanning the range [-1,1] 
      
     
     \deqn{D' = \frac{D}{D_{max} } }{D' = D / Dmax}
       
       where, if D > 0:
       \deqn{%
        D_{max} = \min( p_A p_b, p_a p_B )  %
        }{%
        Dmax = min( p(A)p(b), p(a)p(B) )   %
        } 
        or if D < 0:
        \deqn{%
        D_{max} = \max{ -p_A p_B, -p_a p_b }  %
        }{%
        Dmax = max( -p(A)p(B), -p(a)p(b) )  %
        }
    }
       
     \item{r}{ correlation coefficient between the markers
     
     \deqn{%
     r = \frac{-D}{\sqrt( p_A * p_a * p_B * p_b  )} %
      }{%
      r = -D / sqrt( p(A) * p(a) * p(B) * p(b) ) %
      }
      }
  }
        
   where
   \itemize{
   \item{-}{ \eqn{p_A}{p(A)} is defined as the observed probability of
   allele 'A' for marker 1, }
   \item{-}{ \eqn{p_a=1-p_A}{p(a) = 1-p(A)} is defined as the observed probability of
  allele 'a' for marker 1, }
  \item{-}{\eqn{p_B}{p(B)} is defined as the observed probability of
  allele 'B' for marker 2, and }
  \item{-}{\eqn{p_b=1-p_B}{p(b) = 1- p(B)} is defined as the observed probability of
  allele 'b' for marker 2, and }
   \item{-}{\eqn{p_{AB}}{p(AB)} is defined as the probability of
   the marker allele pair 'AB'. }
   }
   
   For genotype data, AB/ab cannot be distinguished from
   aB/Ab. Consequently, we estimate \eqn{p_{AB}}{p(AB)} using maximum
   likelihood and use this value in the computations.
}
\examples{
# phased data
data('population', package = 'LDtools')
LD(x = X[, 1], y = X[, 2], is_phased = TRUE, any_na = FALSE)

# unphased data
tmp <- transform_geno(.gen_geno(n = 20L, m = 2L))
LD(x = tmp$geno[, 1], y = tmp$geno[, 2], is_phased = FALSE,
any_na = TRUE)

}
\author{
Dominik Mueller (\email{dominikmueller64@yahoo.de})
The documentation is adapted from the \link[genetics]{LD} function of the 
\href{https://cran.r-project.org/web/packages/genetics/index.html}{genetics} 
package by \href{https://sites.google.com/a/warnes.net/resume/}{Gregory Warnes}.
}
\seealso{
{ \code{\link{transform_geno}}, \code{\link{LD_mult}}  }
}

