% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/package_documentation.R
\docType{package}
\name{LDtools-package}
\alias{LDtools}
\alias{LDtools-package}
\title{LDtools: a package for LD measures and plots for detecting selection}
\description{
Computes measures and plots of LD for detecting selection
}
\details{
\tabular{ll}{
 Package: \tab LDtools\cr
 Type: \tab Package\cr
 Version: \tab 0.1\cr
 Date: \tab 2016-06-28\cr
 License: \tab GPL-2\cr
 }
}
\section{LDtools functions}{

The function DPAplot() constructs Disequilibrium Patern Analysis (DPA) plots for haplotype frequency data.
The function lsort() sort a data.frame on specific columns.
}
\references{
xxx
}
\keyword{LD}
\keyword{disequilibrium}
\keyword{genetics}
\keyword{linkage}

