% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDtools.R
\docType{package}
\name{LDtools}
\alias{LDtools}
\alias{LDtools-package}
\title{Computation of linkage disequilibrium from biallelic markers.}
\description{
This provides efficient functions for computing linkage
disequilibrium (LD) between biallelic marker loci from either phased
(haplotype) or unphased (genotype) data.
}
\details{
All performance-critical parts are implemented in C++.
}
\section{Functions}{

}

