% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{map}
\alias{map}
\title{Genetic map}
\format{A data.frame with a the columns: \code{id} gives the unique
identifier of the locus, \code{pos} gives the genetic position and 
\code{is_qtl} indicates if the locus is a QTL.}
\description{
The genetic map of the chromsome.
}
