% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{NIreland.freqs}
\alias{NIreland.freqs}
\title{Pairwise haplotype frequencies for Northern Ireland population}
\format{A data frame with 1317 observations on the following 5 variables.
  \describe{
    \item{\code{locus1}}{a character vector}
    \item{\code{locus2}}{a character vector}
    \item{\code{allele1}}{a character vector}
    \item{\code{allele2}}{a character vector}
    \item{\code{haplo.freq}}{a numeric vector}
  }}
\source{
Williams F, Meenagh A, Single R, McNally M, Kelly P, Nelson MP, Meyer D, Lancaster A, Thomson G, and Middleton D. High resolution HLA-DRB1 identification of a Caucasian population. Human Immunology, 2004; 65(1): 66-77
}
\usage{
data(NIreland.freqs)
}
\description{
HLA haplotype frequencies for 6 pairs of HLA loci from Williams et al. 2004 for 1000 individuals from the Northern
Ireland population.
}
\details{
HLA-A, - B, - C, and -DRB1 loci
}
\references{
http://www.ncbi.nlm.nih.gov/pubmed/14700598
}
\keyword{data}

