% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{population}
\alias{population}
\title{A population for examples in LD computations.}
\format{The matrix \code{X} contains haplotype data for 100 individuals
at 2000 simulated SNPs and 200 QTL. The data.frame \code{map} contains a
genetic map of the chromosome.}
\description{
The data set contains a simulated population
(one chromosome) for illustration LD computations.
}
