% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window.R
\name{sliding_window}
\alias{sliding_window}
\title{Compute summary statistics on a sliding window}
\usage{
sliding_window(x, pos, start, width, advance, stat = c("mean", "median",
  "min", "max", "sd"))
}
\arguments{
\item{x}{A numeric vector containing the values.}

\item{pos}{A incresingly sorted vector containing the corresponding
positions.}

\item{start}{A double. The starting point of the sliding window.}

\item{width}{A double. The width of the sliding window.}

\item{advance}{A double. The step size of the sliding window.}

\item{stat}{A string. The summary statistic to be calculated.}
}
\value{
A \code{data.frame} with columns \code{start} (start of the
window), \code{end} (end of the window), \code{stat} (the computed
summary statistic) and \code{n} (number of data points in the window).
}
\description{
Compute summary statistics on a sliding window along a
vector containing data where the positions of the data points are
stored in another vector.
}
\details{
Note that only windows that fully fit into the range of \code{pos} are
considered.
}
\examples{
set.seed(123L)
n <- 1000L
x <- arima.sim(n = n, list(ar = 0.99))
pos <- sort(runif(n))
plot(x = pos, y = x)
advance <- 0.01
width <- c(0.02, 0.1, 0.2)
colors <- c('red', 'green', 'blue')
for (i in seq_along(width)) {
  df <- sliding_window(x, pos, 0.0, width[i], advance, "mean")
  points(x = df$begin + width[i] / 2, df$stat, col = colors[i], pch = 19, lwd = 2.0, type = 'l')
}
}
\author{
Dominik Mueller (\email{dominikmueller64@yahoo.de}).
}
