% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_geno.R
\name{transform_geno}
\alias{transform_geno}
\title{Transform genotypes}
\usage{
transform_geno(geno, sep = "/", check = TRUE)
}
\arguments{
\item{geno}{A character matrix containing the genotypes.}

\item{sep}{A character string use to split the genotypes in \code{geno}.}

\item{check}{Should checks be performed?}
}
\value{
A numeric matrix of the same dimensions as geno containing the
recoded genotypes.
}
\description{
Transform genotypes from two-letter coding to numeric
coding.
}
\details{
A genotype must consist of a sequence of three characters,
where the mittle character is the seperator as specified in
\code{sep}. Recoding is performed based on the major allele at the
respective locus, i.e., if for instance the major allele is "A",
the recoded genotypes of "A/A", "A/B", "B/A" and "B/B" will be,
2, 1, 1, 0, given \code{sep = "/"}.
}
\examples{
geno <- structure(c("A/G", "G/A", "A/A", NA, "A/A", "G/G", NA, "G/G",
                    "G/G", "A/A", "G/G", NA, "C/C", "C/C", "C/A", NA, "C/C", "C/C"),
                 .Dim = c(6L, 3L))
transform_geno(geno, sep = '/')

}
\author{
Dominik Mueller (\email{dominikmueller64@yahoo.de})
}
