% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{X}
\alias{X}
\title{Haplotype data}
\format{A matrix with haplotype data for 100 diploid individuals at
2000 simulated SNPs and 200 QTL. The alleles are coded with 0 and 1.}
\description{
The haplotype data of the simulated population
}

